requirements.txt
Django>=2.2,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2~=2.9.5
drf-spectacular>=0.22.1,<0.23
django-cors-headers>=3.7.0,<4.1.0
django-celery-beat==2.2.1
django-celery-results==2.1.0
django-ninja
psycopg2-binary==2.9.1
pymongo==4.1.1
celery==5.2.0
redis==3.5.3
Pillow>=9.1.0,<9.2